Rename `hy._compat` to `hy.compat`
