+++ hy/compat.py
