+++ tests/compilers/test_ast.py
@@ -4,7 +4,7 @@ from textwrap import dedent
 import pytest
 
 import hy
-from hy._compat import PY3_11
+from hy.compat import PY3_11
 from hy.compiler import hy_compile
 from hy.errors import HyError, HyLanguageError
 from hy.reader import read_many
