+++ tests/native_tests/deftype.hy
@@ -1,4 +1,4 @@
-(do-mac (when hy._compat.PY3_12 '(do
+(do-mac (when hy.compat.PY3_12 '(do
 
 (import tests.resources.tp :as ttp)
